# 02-development/homework/balancio/openapi.yaml
openapi: 3.0.3
info:
  title: Balancio API
  description: >
    Shared expense splitter for friends, households, couples, trips, and
    other small groups. See _docs/specs.md for the full product spec.
  version: "0.1.0"
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /expenses:
    get:
      tags: [expenses]
      summary: List expenses visible to the current user
      parameters:
        - name: group_id
          in: query
          schema:
            type: string
        - name: member_id
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: search
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: integer
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of expenses, newest expense_date first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpensePage"
    post:
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /expenses/{expense_id}:
    parameters:
      - $ref: "#/components/parameters/ExpenseId"
    get:
      tags: [expenses]
      summary: Get an expense by id
      responses:
        "200":
          description: The expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [expenses]
      summary: Update an expense (creator or admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: >
            Validation error, or the edit is blocked because confirmed
            settlement activity depends on this expense's current balance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [expenses]
      summary: Delete an expense permanently (creator or admin only)
      responses:
        "204":
          description: Expense deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /expenses/{expense_id}/duplicate:
    parameters:
      - $ref: "#/components/parameters/ExpenseId"
    post:
      tags: [expenses]
      summary: >
        Get a draft copy of an expense for review (title/category/participants/
        note/tags copied; date and payer amounts reset). Does not persist
        anything until the draft is submitted via POST /expenses.
      responses:
        "200":
          description: Draft expense input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseInput"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ExpenseId:
      name: expense_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Validation error or invalid financial state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Friendly, domain-specific message. Never a stack trace.
...
    Money:
      type: string
      description: Decimal amount with exactly two decimal places.
      pattern: '^\d+\.\d{2}$'
      example: "123.45"
...
    Payer:
      type: object
      required: [user_id, amount]
      properties:
        user_id:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
    Share:
      type: object
      required: [user_id, amount]
      properties:
        user_id:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
    Expense:
      type: object
      required:
        - id
        - group_id
        - title
        - amount
        - expense_date
        - category_id
        - tags
        - created_by
        - payers
        - participant_ids
        - shares
        - created_at
        - updated_at
      properties:
        id:
          type: string
        group_id:
          type: string
        title:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        expense_date:
          type: string
          format: date
        category_id:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        created_by:
          type: string
        payers:
          type: array
          items:
            $ref: "#/components/schemas/Payer"
        participant_ids:
          type: array
          items:
            type: string
        shares:
          type: array
          items:
            $ref: "#/components/schemas/Share"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required:
        - group_id
        - title
        - amount
        - expense_date
        - payers
        - participant_ids
      properties:
        group_id:
          type: string
        title:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        expense_date:
          type: string
          format: date
        category_id:
          type: string
          nullable: true
        note:
          type: string
        tags:
          type: array
          items:
            type: string
          default: []
        payers:
          type: array
          items:
            $ref: "#/components/schemas/Payer"
          description: Amounts must sum exactly to `amount`.
        participant_ids:
          type: array
          items:
            type: string
          description: Equal shares are computed by the server.
    ExpenseUpdate:
      type: object
      description: All fields optional; only provided fields are changed.
      properties:
        title:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        expense_date:
          type: string
          format: date
        category_id:
          type: string
          nullable: true
        note:
          type: string
        tags:
          type: array
          items:
            type: string
        payers:
          type: array
          items:
            $ref: "#/components/schemas/Payer"
        participant_ids:
          type: array
          items:
            type: string
    ExpensePage:
      type: object
      required: [items, next_cursor, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        next_cursor:
          type: integer
          nullable: true
        total:
          type: integer
...